# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Приют API",
    "summary": "API mini-app для hospitality exchange в MAX",
    "description": "Профили, предложения жилья, запросы, отзывы и уведомления. Кроме health endpoints, API требует HttpOnly сессию, полученную после проверки initData MAX. Изменяющие запросы также проверяют Origin.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://64.188.79.42",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/localities/suggest": {
      "get": {
        "tags": [
          "localities"
        ],
        "summary": "Suggest",
        "operationId": "suggest_api_localities_suggest_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 120,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocalityResponse"
                  },
                  "title": "Response Suggest Api Localities Suggest Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/localities/{locality_id}": {
      "get": {
        "tags": [
          "localities"
        ],
        "summary": "Read Locality",
        "operationId": "read_locality_api_localities__locality_id__get",
        "parameters": [
          {
            "name": "locality_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Locality Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalityResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/me/listing": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "Read My Listing",
        "operationId": "read_my_listing_api_me_listing_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "listings"
        ],
        "summary": "Put My Listing",
        "description": "Create the user's single offer, or replace its editable fields.",
        "operationId": "put_my_listing_api_me_listing_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "listings"
        ],
        "summary": "Remove My Listing",
        "description": "Hard-delete the offer. Any stay_requests against it are cascaded away\nat the database level — the host is warned about this client-side.",
        "operationId": "remove_my_listing_api_me_listing_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/listings": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "List Listings",
        "operationId": "list_listings_api_listings_get",
        "parameters": [
          {
            "name": "locality_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Locality Id"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "guests",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 8,
              "minimum": 1,
              "default": 1,
              "title": "Guests"
            }
          },
          {
            "name": "accommodation_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(room|apartment|house|sofa)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accommodation Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingResponse"
                  },
                  "title": "Response List Listings Api Listings Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/listings/{listing_id}": {
      "get": {
        "tags": [
          "listings"
        ],
        "summary": "Read Listing",
        "operationId": "read_listing_api_listings__listing_id__get",
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Listing Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HostSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "locality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalityResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "bio": {
            "type": "string",
            "title": "Bio"
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Interests"
          },
          "avatarEmoji": {
            "type": "string",
            "title": "Avataremoji"
          },
          "avatarColor": {
            "type": "string",
            "title": "Avatarcolor"
          },
          "photoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photourl"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "city",
          "locality",
          "bio",
          "interests",
          "avatarEmoji",
          "avatarColor",
          "photoUrl"
        ],
        "title": "HostSummary"
      },
      "ListingInput": {
        "properties": {
          "localityId": {
            "type": "string",
            "format": "uuid",
            "title": "Localityid"
          },
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "shortDescription": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Shortdescription"
          },
          "description": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Description"
          },
          "guests": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Guests"
          },
          "accommodationType": {
            "type": "string",
            "pattern": "^(room|apartment|house|sofa)$",
            "title": "Accommodationtype"
          },
          "availableFrom": {
            "type": "string",
            "format": "date",
            "title": "Availablefrom"
          },
          "availableTo": {
            "type": "string",
            "format": "date",
            "title": "Availableto"
          },
          "tags": {
            "items": {
              "type": "string",
              "maxLength": 80,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 12,
            "title": "Tags"
          },
          "amenities": {
            "items": {
              "type": "string",
              "maxLength": 80,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 20,
            "title": "Amenities"
          },
          "rules": {
            "items": {
              "type": "string",
              "maxLength": 80,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 20,
            "title": "Rules"
          },
          "photoUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 700000
              },
              {
                "type": "null"
              }
            ],
            "title": "Photourl"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "localityId",
          "title",
          "shortDescription",
          "description",
          "guests",
          "accommodationType",
          "availableFrom",
          "availableTo"
        ],
        "title": "ListingInput"
      },
      "ListingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "host": {
            "$ref": "#/components/schemas/HostSummary"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "localityId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Localityid"
          },
          "locality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocalityResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "shortDescription": {
            "type": "string",
            "title": "Shortdescription"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "guests": {
            "type": "integer",
            "title": "Guests"
          },
          "accommodationType": {
            "type": "string",
            "title": "Accommodationtype"
          },
          "availableFrom": {
            "type": "string",
            "format": "date",
            "title": "Availablefrom"
          },
          "availableTo": {
            "type": "string",
            "format": "date",
            "title": "Availableto"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "amenities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Amenities"
          },
          "rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rules"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "reviewsCount": {
            "type": "integer",
            "title": "Reviewscount"
          },
          "photos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photos"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "host",
          "city",
          "localityId",
          "locality",
          "title",
          "shortDescription",
          "description",
          "guests",
          "accommodationType",
          "availableFrom",
          "availableTo",
          "tags",
          "amenities",
          "rules",
          "rating",
          "reviewsCount",
          "photos"
        ],
        "title": "ListingResponse"
      },
      "LocalityResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "typeShort": {
            "type": "string",
            "title": "Typeshort"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "shortLabel": {
            "type": "string",
            "title": "Shortlabel"
          },
          "fullLabel": {
            "type": "string",
            "title": "Fulllabel"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "typeShort",
          "region",
          "district",
          "shortLabel",
          "fullLabel"
        ],
        "title": "LocalityResponse"
      },
...
      "MyListingResponse": {
        "properties": {
          "listing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListingResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "listing"
        ],
        "title": "MyListingResponse"
      },
...
    }
  }
}